\name{singlyLinkedList-package}
\alias{singlyLinkedList-package}
\alias{singlyLinkedList}
\docType{package}
\title{
What the package does (short line)
Singly Linked List
}
\description{
A Singly Linked List
}
\details{
\tabular{ll}{
Package: \tab singlyLinkedList\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-11-10\cr
License: \tab MIT\cr
Depends: \tab methods\cr
}
}
\author{
Dan Nuffer
}
\keyword{ package }
\examples{
list = new("singlyLinkedList")
}
